Final Project 3 Submission
Admittedly, my commits are more cosmetic than functional.  I had
originally started Project 3 on a Windows PC and had many commits
regarding functionality and setting up the JS files as well as the
initial editing of the index file and other pages from Project 2 to put
everything on the same page.  However, Git for Windows would not sync
with Github.  It was stuck at "syncing...0%."  I left it running
overnight and attempted to re-sync multiple times but nothing worked.
Switched the completed files over to the Mac and voila Git for Mac
synced up perfectly the first time.  My commits were lost so they
aren't as extensive as they previously were but I did make some
cosmetic changes to the overall feel of the site to improve the user
experience.